URL justification
Reinforce HTTPS only.
