+++ assets/layouts/styles.css
@@ -180,12 +180,14 @@ end */
   border-bottom-left-radius: 2px;
   border-bottom-right-radius: 2px;
 }
+/* mod: with our default img */
 .demo-blog .on-the-road-again .mdl-card__media {
-  background-image: url('images/road.jpg');
+  background-image: url('/assets/images/default.jpg');
 }
 .demo-blog .shopping .mdl-card__media {
-  background-image: url('images/shopping.jpg');
+  background-image: url('/assets/images/default.jpg');
 }
+/* end */
 .demo-blog .demo-blog__posts > .demo-nav {
   margin: 12px 15px;
   display: flex;
