+++ templ/entry.html
@@ -42,8 +42,7 @@ article blockquote:before, article blockquote:after { content: none; }
     </div>
     <div class="demo-blog__posts mdl-grid">
       <div class="mdl-card mdl-shadow--4dp mdl-cell mdl-cell--12-col">
-        <div class="mdl-card__media mdl-color-text--grey-50" style="background: {{if $.A.Background | bgImg -}}url('{{$.A.Background | cdn}}')
-             {{- else -}}{{$.A.Background}}{{- end}};">
+        <div class="mdl-card__media mdl-color-text--grey-50" banner="{{$.A.Background}}">
           <h3>{{$.A.Title}}</h3>
         </div>
         <div class="mdl-color-text--grey-700 mdl-card__supporting-text meta">
@@ -136,7 +135,8 @@ article blockquote:before, article blockquote:after { content: none; }
     }
   }
 
-  styleCover(document.getElementsByClassName('mdl-card__media')[0]);
+  var el = document.getElementsByClassName('mdl-card__media')[0];
+  banner(el.getAttribute('banner'), el);
 
   document.getElementById('home').addEventListener('click', function () {
     window.open(window.location.protocol + '//' + window.location.host, "_self");
